feat(303): comments popover for mobile
